refactor: move temp-file serving helpers into utils
_remove_file / _send_generated_file were private to the API blueprint, but
the UI wordlist download needs the same "generate into control/tmp, serve it,
unlink after the response" behaviour. Move them to hashview.utils.utils as
remove_file / send_generated_file so both blueprints share one implementation
rather than the UI growing a second copy of the direct_passthrough workaround.

No behaviour change; the two helper unit tests now target utils.
